First version of the PoG winner selection algorithm.
This version constructs a discrete distribution using the ANVs of the wallets
and uses Inverse Transform Sampling to sample from the distribution.

A WalletSelected can return N winners by sampling the distribution based
on the input hash. The sampling is deterministic and returns the same winners
given an input hash.

This version samples in memory.