Wrote function to get all ANVs from DB into memory.
Added a function to ReferralsViewDB to get all ANVs into memory.
The vector of KeyANVs is used by the selection algorithm to choose N winners.

In the future we should store a heap on disk of KeyANVs and change the selection
algorithm to use the disk instead. Alternatively we can store a heap as an mmap
file and change the selection algorithm to index using pointer arithmatic.