Fix ANV computation and referral network when reindexing the blockchain.
CheckReferral function was supposed to be context less to validate
a referral outside of other referrals. However it was written using
the referral db and therefore the contextless assumption was broken.

I removed the context based check and the blockchain was able to be reindexed
and loaded correctly. Also I made sure to index the genesis referral when
ConnectBlock was used on the genesis block.