Creating a new type of script called a paramed pay-to-script-hash
This is similar to a normal pay-to-script-hash except that you can
add params at the end which are then put in the top of the stack in
the deserialized script.

Looks like

    OP_HASH160 <script hash> OP_EQUALVERIFY <param a> <param b> <...>

Whichever script is hashed to the script hash will have params a,b, etc on
top of the stack to use. This way you can pull out specific parameters
and have a scripthash that doesn't change unless params change.