Fix the ANV computation and getanv RPC call.
The ANV computation incorrectly debited ANVs because it called
AccessCoin after the coin was spent. The coins view only has unspent
transactions. Since UpdateCoins spends the coins in the view, it isn't
good to call AccessCoin on coins that are expected to be unspent. The fix
gathers the debits and credits first, and then modifies the ANV after
the block is procesed. This ensures we do the correct thing.

This commit also provides script id support to the getanv RPC call.