Started work on a more efficient lottery system.
Will use a weighted random sampling algorithm by
Efraimidis and Spirakis. We will have a reservoir of X addresses
stored with a weighted key.

As we mine blocks we will insert the addresses into the reservoir with
some probability using a uniform random number taken to the power of the
1 over the weight. Our implementation will use the log version of that
for efficiency reasons.

A min-heap will have to be implemented on top of the DB. It is still up in
the air if using a separate storage system for the reservoir makes sense.