Validate transaction and invite inputs.
Transaction and invite inputs must be of the same kind. Validation has
been added to query all inputs and make sure the transactions are the
same kind. If the transaction is a normal transaction, then the inputs
must be normal transactions. If the transaction is an invite, then the inputs
must be invites. We cannot cross the streams.