Don't update model in realtime even after initial sync is false.
The model would update per block instead of every 250ms if initial sync is false.
However, The initial sync can be false up to 5 hours (block time) before it is actually synced up.
Because syncing could be must faster than 4 blocks a second this would cause the
UI to struggle re-rendering if the blocks came in too fast slowing the whole
UI down.