Fix use of invalidated iterator which can cause a crash.
A crash can happen when using an invalidated iterator when an orphan referral
is removed from the mempool. This commit fixes that issue by erasing
the iterator at the end after using it.

Another potential source of crashes is the EraseOrphanfor function which
iterated over the transaction and referral maps while erasing from the maps.
The code was changed to step over the potentially erased iterator.