CheckAddressBeacon fails to give correct return value.
CheckAddressBeacon uses the in memory cache to figure out
if an address has already been beaconed. This cache wasn't properly
updated when DisconnectBlock was called. This means if a referral
was beaconed on an orphaned branch of the blockchain, All nodes will reject all
future blocks with that beacon causing mini forks and bans throughout the network.