Fix confusing transition between reindexing and downloading.
During Reindexing, new headers are not downloaded until reindexing is
complete. This causes a strange behavior where the progress bar jumps back down
because the system recognized there are new headers to download.

Now the dialog clearly states when reindexing is finished and when downloading
begins.

Also downloading progress is now shown relative to when you started Merit.
So instead of showing usually 99% and up, it shows progress of downloading
remaining blocks from start to end.