Added old and new distribution.
1. renamed anv to cgs.
2. the address selector can now sample from two distributions, one base
on beacon age.

TODO

Need to get beacon age in an efficient way. Plan will be to create a
lazy on demand index, that will store age. If age isn't in the index it
will search for block hash and use that. Because of the way the sampling
works, most addresses should be indexed this way in the very first block
that is activated.