Work on creating new/old ambassador rewards.
There are two slots, the old slots and the new slots. The old
slots are reserved for older beacons and they get the majority of
ambassador rewards. The new slots are fewer and provide rewards to new
ambassadors. This balances fairness with new people experiencing the
benefits of ambassador rewards.